Add better URL parameter handler for Youtube
The interface was assuming that the  last parameter of the
Youtube URL was the video id, sometimes this is not true. For
example when you paste a URL from a youtube list, the URL has
extra arguments (like list and index), in this scenario taking
the last argument does not work.

With this patch you can paste any valid youtube URL and it will be
played.